Update function docbloc
